requirements.txt
matplotlib==3.10.3
numpy==2.3.1
pandas==2.3.1
Pillow==11.3.0
scikit_learn==1.7.0
seaborn==0.13.2
streamlit==1.46.1
tensorflow==2.19.0
